Use DBus activation for applications that are dbus activatable
Use dbus activation as per freedesktop desktop entry spec. Applications that
want to use it need to have DBusActivatable=true in their desktop file, have a
dbus-activatable service with the same name as their desktop file and implement
the org.freedesktop.Application interface at the object path that results when
every '.' in the name of the service is replaced by '/'. KDBusService
automatically takes care of registering the name and object that implements
the interface. The application just needs to connect to activate(Action)Requested
signals of KDBusService.